Add module "global", move globalDB in "global"
