+++ internal/global/global.go
@@ -1,0 +1,7 @@
+package global
+
+import (
+	"go_project/internal/models"
+)
+
+var DB = make(map[int]models.User)
